backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.2
python-multipart==0.0.17
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
python-dateutil==2.9.0.post0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
alembic==1.14.0
pytest==8.3.4
httpx==0.27.2
